Move inference dataset input features tests to test_inference.py
